Fixes math in SimulationState SequenceTimeOfNextCapture
SequenceTimeOfNextCapture was reporting time with respect to two
different frames depending on context: either it would report absolute
time since the start of the sequence, if the first capture had not been
completed, or it would report the amount of time left until the next
capture, relative to the current time.

Removed logic which reported relative time, which seems like it was not
part of any code path, or was being corrected for elsewhere in
SimulationState's update loop. Removed some of the error correction in
the update loop and added stricter expectations for the management of
the timing code.

Added unit tests to ensure function reports time we'd expect regardless
of context.

Created follow-up task AISV-845 for making the timing code less fragile.